feat: enable live STT to everyone
