--- front/hooks/useVoiceTranscriberService.ts
@@ -1,535 +1,0 @@
-/// <reference types="chrome" />
-import type { FileUploaderService } from "@app/hooks/useFileUploaderService";
-import { useSendNotification } from "@app/hooks/useNotification";
-import {
-  hasWebkitAudioContext,
-  quackingVoiceTranscriptService,
-  startLevelMeteringInterval,
-  useElapsedSeconds,
-  type VoiceTranscriberService,
-  type VoiceTranscriberStatus,
-} from "@app/hooks/utils/voice";
-import { clientFetch } from "@app/lib/egress/client";
-import {
-  isBrowserExtension,
-  isChromeExtension,
-  isFirefoxExtension,
-} from "@app/lib/utils/extension";
-
-import type { AugmentedMessage } from "@app/lib/utils/find_agents_in_message";
-import { normalizeError } from "@app/types/shared/utils/error_utils";
-import type { LightWorkspaceType } from "@app/types/user";
-import type { MutableRefObject } from "react";
-import { useCallback, useEffect, useRef, useState } from "react";
-
-// We are using webm with Opus codec
-// In general browsers are using a 48 kbps bitrate
-// A 3-minute recording will be around 1200kB
-// 180 seconds * 48000bps / 8 => 1 080 000 bit, round up to 1200kB
-const MAXIMUM_FILE_SIZE_FOR_INPUT_BAR_IN_BYTES = 1200 * 1024;
-
-const MICROPHONE_POPUP_TIMEOUT_MS = 60_000; // 1 minute
-
-interface UseVoiceTranscriberServiceParams {
-  owner: LightWorkspaceType;
-  onTranscribeDelta?: (delta: string) => void;
-  onTranscribeComplete?: (transcript: AugmentedMessage[]) => void;
-  onError?: (error: Error) => void;
-  fileUploaderService: FileUploaderService;
-}
-
-export function useVoiceTranscriberService({
-  owner,
-  onTranscribeDelta,
-  onTranscribeComplete,
-  onError,
-  fileUploaderService,
-}: UseVoiceTranscriberServiceParams): VoiceTranscriberService {
-  const [status, setStatus] = useState<VoiceTranscriberStatus>("idle");
-  const [level, setLevel] = useState(0);
-  const elapsedSeconds = useElapsedSeconds(status);
-
-  const mediaRecorderRef = useRef<MediaRecorder | null>(null);
-  const chunksRef = useRef<Blob[]>([]);
-  const streamRef = useRef<MediaStream | null>(null);
-  const audioContextRef = useRef<AudioContext | null>(null);
-  const analyserRef = useRef<AnalyserNode | null>(null);
-  const sourceNodeRef = useRef<MediaStreamAudioSourceNode | null>(null);
-  const intervalRef = useRef<NodeJS.Timeout | null>(null);
-
-  const sendNotification = useSendNotification();
-
-  const stopLevelMetering = useCallback(() => {
-    if (intervalRef.current !== null) {
-      clearInterval(intervalRef.current);
-      intervalRef.current = null;
-    }
-
-    try {
-      analyserRef.current = null;
-      sourceNodeRef.current?.disconnect();
-      sourceNodeRef.current = null;
-      if (audioContextRef.current) {
-        // Close the AudioContext to release resources.
-        void audioContextRef.current.close();
-      }
-    } catch {
-      // Ignore errors on cleanup.
-    } finally {
-      audioContextRef.current = null;
-      setLevel(0);
-      setStatus("idle");
-    }
-  }, []);
-
-  useEffect(() => {
-    return () => {
-      stopLevelMetering();
-    };
-  }, [stopLevelMetering]);
-
-  const startLevelMetering = useCallback(
-    (stream: MediaStream) => {
-      stopLevelMetering();
-      try {
-        const AC = hasWebkitAudioContext(window)
-          ? window.webkitAudioContext
-          : window.AudioContext;
-        const ctx = new AC();
-        audioContextRef.current = ctx;
-        const analyser = ctx.createAnalyser();
-        analyser.fftSize = 2048;
-        analyser.smoothingTimeConstant = 0.85;
-        const source = ctx.createMediaStreamSource(stream);
-        source.connect(analyser);
-        analyserRef.current = analyser;
-        sourceNodeRef.current = source;
-
-        intervalRef.current = startLevelMeteringInterval(
-          analyser,
-          analyserRef,
-          setLevel
-        );
-      } catch {
-        // If metering fails (unsupported), we silently ignore.
-        audioContextRef.current = null;
-        analyserRef.current = null;
-      }
-    },
-    [stopLevelMetering]
-  );
-
-  // Cleanup on unmounting: ensure the recorder is stopped and tracks are closed.
-  useEffect(() => {
-    return () => {
-      stopRecorder(mediaRecorderRef.current);
-      stopLevelMetering();
-      stopTracks(streamRef.current);
-    };
-  }, [stopLevelMetering]);
-
-  const startRecording = useCallback(async () => {
-    if (status === "recording" || status === "authorizing_microphone") {
-      return;
-    }
-    try {
-      setStatus("authorizing_microphone");
-      const stream = await requestMicrophone();
-      streamRef.current = stream;
-
-      const recorder = createRecorder(stream, chunksRef);
-      mediaRecorderRef.current = recorder;
-
-      // Start level metering alongside recording.
-      startLevelMetering(stream);
-
-      recorder.start();
-
-      setStatus("recording");
-    } catch (err) {
-      const isPermissionError =
-        err instanceof DOMException &&
-        (err.name === "NotAllowedError" ||
-          err.name === "PermissionDeniedError");
-
-      sendNotification({
-        type: "error",
-        title: isPermissionError
-          ? "Microphone permission required."
-          : "Could not start recording.",
-        description: isPermissionError
-          ? "Please allow microphone access and try again."
-          : "Your browser may not support voice recording.",
-      });
-      setStatus("idle");
-    }
-  }, [sendNotification, startLevelMetering, status]);
-
-  const finalizeRecordingTranscribeToInputBar = useCallback(
-    async (file: File) => {
-      const form = new FormData();
-      form.append("file", file);
-
-      const resp = await clientFetch(
-        `/api/w/${owner.sId}/services/transcribe`,
-        {
-          method: "POST",
-          body: form,
-        }
-      );
-
-      if (!resp.ok) {
-        const msg = await resp.text();
-        sendNotification({
-          type: "error",
-          title: "Voice upload failed.",
-          description: msg || "Failed to send audio for transcription.",
-        });
-        return;
-      }
-
-      // Stream Server-Sent Events and forward deltas.
-      const body = resp.body;
-      if (!body) {
-        sendNotification({
-          type: "error",
-          title: "Transcription failed.",
-          description: "Empty response while streaming transcription.",
-        });
-        return;
-      }
-
-      await readSSEFromPostRequest({
-        body,
-        onTranscribeDelta,
-        onTranscribeComplete,
-        onError,
-      });
-    },
-    [
-      onTranscribeDelta,
-      onTranscribeComplete,
-      onError,
-      owner.sId,
-      sendNotification,
-    ]
-  );
-
-  const finalizeRecordingAddAsAttachment = useCallback(
-    async (file: File) => {
-      await fileUploaderService.handleFilesUpload([file]);
-    },
-    [fileUploaderService]
-  );
-
-  const stopAndFinalize = useCallback(async () => {
-    setStatus("transcribing");
-
-    try {
-      const mimeType = mediaRecorderRef.current?.mimeType || "audio/webm";
-      const file = buildAudioFile(chunksRef.current, mimeType);
-      chunksRef.current = [];
-
-      if (file.size <= MAXIMUM_FILE_SIZE_FOR_INPUT_BAR_IN_BYTES) {
-        await finalizeRecordingTranscribeToInputBar(file);
-      } else {
-        await finalizeRecordingAddAsAttachment(file);
-      }
-    } catch (e) {
-      sendNotification({
-        type: "error",
-        title: "Recording error.",
-        description:
-          e instanceof Error ? e.message : "An unknown error occurred.",
-      });
-    } finally {
-      stopLevelMetering();
-      stopTracks(streamRef.current);
-      streamRef.current = null;
-    }
-  }, [
-    finalizeRecordingTranscribeToInputBar,
-    finalizeRecordingAddAsAttachment,
-    sendNotification,
-    stopLevelMetering,
-  ]);
-
-  const stopRecording = useCallback(async () => {
-    const recorder = mediaRecorderRef.current;
-    if (!recorder || recorder.state === "inactive") {
-      return;
-    }
-
-    const stopped = new Promise<void>((resolve) => {
-      recorder.onstop = () => resolve();
-    });
-
-    recorder.stop();
-    await stopped;
-    await stopAndFinalize();
-  }, [stopAndFinalize]);
-
-  return owner.metadata?.allowVoiceTranscription !== false
-    ? {
-        status,
-        level,
-        elapsedSeconds,
-        startRecording,
-        stopRecording,
-      }
-    : quackingVoiceTranscriptService;
-}
-
-// Helpers ---------------------------------------------------------------------
-
-const stopTracks = (stream: MediaStream | null) => {
-  if (!stream) {
-    return;
-  }
-  stream.getTracks().forEach((t) => t.stop());
-};
-
-const stopRecorder = (recorder: MediaRecorder | null) => {
-  if (!recorder) {
-    return;
-  }
-  if (recorder.state !== "inactive") {
-    recorder.stop();
-  }
-};
-
-const buildAudioFile = (chunks: Blob[], mimeType: string) => {
-  const blob = new Blob(chunks, { type: mimeType });
-  const ext = mimeType.includes("mp4") ? "mp4" : "webm";
-  const filename = `voice-${new Date().toISOString()}.${ext}`;
-
-  return new File([blob], filename, { type: mimeType });
-};
-
-const getMicrophonePermissionState = async (): Promise<PermissionState> => {
-  const status = await navigator.permissions.query({
-    name: "microphone",
-  });
-  return status.state;
-};
-
-const openMicrophoneAccessPopup = async (): Promise<void> => {
-  const createdWindow = await chrome.windows.create({
-    url: chrome.runtime.getURL("request-mic.html"),
-    type: "popup",
-    width: 400,
-    height: 400,
-  });
-
-  if (!createdWindow?.id) {
-    return;
-  }
-
-  const windowId = createdWindow.id;
-
-  await new Promise<void>((resolve, reject) => {
-    const timeout = setTimeout(() => {
-      chrome.windows.onRemoved.removeListener(onWindowClose);
-      chrome.windows.remove(windowId);
-      reject(new Error("Microphone access popup timed out"));
-    }, MICROPHONE_POPUP_TIMEOUT_MS);
-
-    function onWindowClose(closedWindowId: number) {
-      if (closedWindowId === windowId) {
-        clearTimeout(timeout);
-        chrome.windows.onRemoved.removeListener(onWindowClose);
-        resolve();
-      }
-    }
-
-    chrome.windows.onRemoved.addListener(onWindowClose);
-  });
-};
-
-const extensionRequestMicrophonePermission = async (): Promise<MediaStream> => {
-  const state = await getMicrophonePermissionState();
-
-  if (state === "denied") {
-    if (isChromeExtension()) {
-      // Open Chrome extension settings so user can manually re-enable microphone.
-      await chrome.tabs.create({
-        url: `chrome://settings/content/siteDetails?site=chrome-extension%3A%2F%2F${chrome.runtime.id}%2F`,
-      });
-    }
-    // Firefox has no deep link to extension site settings
-    throw new DOMException("Microphone permission denied", "NotAllowedError");
-  }
-
-  if (state === "prompt") {
-    // Side panel can't show the prompt — fall back to popup window
-    await openMicrophoneAccessPopup();
-    const newState = await getMicrophonePermissionState();
-    if (newState !== "granted") {
-      throw new DOMException(
-        "Microphone permission not granted",
-        "NotAllowedError"
-      );
-    }
-    return navigator.mediaDevices.getUserMedia({ audio: true });
-  }
-  // Permission is already granted
-  try {
-    return await navigator.mediaDevices.getUserMedia({ audio: true });
-  } catch (error) {
-    // Firefox bug: when the microphone permission is set to "Always Ask",
-    // navigator.permissions.query incorrectly reports the state as "granted"
-    // instead of "prompt". This causes getUserMedia to fail in the side panel
-    // (which cannot display the native permission prompt). We fall back to the
-    // popup flow so the user can still grant access.
-    if (
-      isFirefoxExtension() &&
-      error instanceof DOMException &&
-      error.name === "NotAllowedError"
-    ) {
-      await openMicrophoneAccessPopup();
-      const newState = await getMicrophonePermissionState();
-      if (newState !== "granted") {
-        throw new DOMException(
-          "Microphone permission not granted",
-          "NotAllowedError"
-        );
-      }
-      return navigator.mediaDevices.getUserMedia({ audio: true });
-    }
-    throw error;
-  }
-};
-
-export const requestMicrophone = async (): Promise<MediaStream> => {
-  if (isBrowserExtension()) {
-    return extensionRequestMicrophonePermission();
-  }
-  return navigator.mediaDevices.getUserMedia({
-    audio: {
-      noiseSuppression: true, // filter out background noise
-      echoCancellation: true, // prevent mic from picking up speaker output
-      autoGainControl: true, // normalize volume
-    },
-  });
-};
-
-// Safari/iOS does not support audio/webm — fall back to audio/mp4.
-const PREFERRED_MIME_TYPES = ["audio/webm;codecs=opus", "audio/mp4"];
-
-function getSupportedMimeType(): string {
-  for (const mime of PREFERRED_MIME_TYPES) {
-    if (MediaRecorder.isTypeSupported(mime)) {
-      return mime;
-    }
-  }
-  // Let the browser pick a default.
-  return "";
-}
-
-const createRecorder = (
-  stream: MediaStream,
-  chunksRef: MutableRefObject<Blob[]>
-): MediaRecorder => {
-  const mimeType = getSupportedMimeType();
-  const recorder = mimeType
-    ? new MediaRecorder(stream, { mimeType })
-    : new MediaRecorder(stream);
-  chunksRef.current = [];
-  recorder.ondataavailable = (e) => {
-    if (e.data && e.data.size > 0) {
-      chunksRef.current.push(e.data);
-    }
-  };
-  return recorder;
-};
-
-// There is no built-in SSE support in the Fetch API for POST, so we manually parse the stream.
-const readSSEFromPostRequest = async ({
-  body,
-  onTranscribeDelta,
-  onTranscribeComplete,
-  onError,
-}: {
-  body: ReadableStream<Uint8Array>;
-  onTranscribeDelta?: (delta: string) => void;
-  onTranscribeComplete?: (transcript: AugmentedMessage[]) => void;
-  onError?: (error: Error) => void;
-}) => {
-  const reader = body.getReader();
-  const decoder = new TextDecoder();
-  let buffer = "";
-  let doneStreaming = false;
-
-  try {
-    // Read until the server signals completion with a \n\n separator after a done event
-    // or the stream ends.
-    while (!doneStreaming) {
-      const { value, done } = await reader.read();
-      if (done) {
-        break;
-      }
-      buffer += decoder.decode(value, { stream: true });
-
-      // SSE frames are separated by double newlines. Support both \n\n and \r\n\r\n.
-      // Process as many complete events as available.
-      while (!doneStreaming) {
-        const nn = buffer.indexOf("\n\n");
-        const rr = buffer.indexOf("\r\n\r\n");
-        if (nn === -1 && rr === -1) {
-          break;
-        }
-        const useNn = nn !== -1 && (rr === -1 || nn < rr);
-        const sep = useNn ? nn : rr;
-        const delimLen = useNn ? 2 : 4;
-        const eventChunk = buffer.slice(0, sep);
-        buffer = buffer.slice(sep + delimLen);
-
-        // Extract the data line(s). Our server sends single-line data payloads.
-        const lines = eventChunk.split(/\r?\n/);
-        for (const line of lines) {
-          const trimmed = line.trim();
-          if (!trimmed) {
-            continue;
-          }
-          if (!trimmed.startsWith("data:")) {
-            continue;
-          }
-          const payload = trimmed.slice(5).trim();
-          if (payload === "done") {
-            doneStreaming = true;
-            break;
-          }
-          try {
-            const parsed = JSON.parse(payload) as
-              | { type: "delta"; delta: string }
-              | { type: "fullTranscript"; fullTranscript: AugmentedMessage[] }
-              | { type: "error"; error: string };
-
-            if (parsed.type === "delta") {
-              // eslint-disable-next-line no-unused-expressions
-              onTranscribeDelta && onTranscribeDelta(parsed.delta);
-            } else if (parsed.type === "fullTranscript") {
-              // eslint-disable-next-line no-unused-expressions
-              onTranscribeComplete &&
-                onTranscribeComplete(parsed.fullTranscript);
-
-              doneStreaming = true;
-            } else if (parsed.type === "error") {
-              // eslint-disable-next-line no-unused-expressions
-              onError && onError(normalizeError(parsed.error));
-              doneStreaming = true;
-            }
-          } catch {
-            // Ignore malformed payloads and continue.
-          }
-        }
-      }
-    }
-  } finally {
-    try {
-      reader.releaseLock();
-    } catch {
-      // ignore
-    }
-  }
-};
