+++ front/hooks/utils/voice.ts
@@ -1,7 +1,15 @@
+/// <reference types="chrome" />
+import {
+  isBrowserExtension,
+  isChromeExtension,
+  isFirefoxExtension,
+} from "@app/lib/utils/extension";
 import type { MutableRefObject } from "react";
 import { useEffect, useState } from "react";
 
 export const SAMPLE_RATE_HZ = 16_000;
+
+const MICROPHONE_POPUP_TIMEOUT_MS = 60_000; // 1 minute
 
 export type VoiceTranscriberStatus =
   | "idle"
@@ -72,3 +80,110 @@ export function startLevelMeteringInterval(
     setLevel(Math.max(0, Math.min(1, (rms - 0.02) / 0.3)));
   }, 250);
 }
+
+export const requestMicrophone = async (): Promise<MediaStream> => {
+  if (isBrowserExtension()) {
+    return extensionRequestMicrophonePermission();
+  }
+  return navigator.mediaDevices.getUserMedia({
+    audio: {
+      noiseSuppression: true, // filter out background noise
+      echoCancellation: true, // prevent mic from picking up speaker output
+      autoGainControl: true, // normalize volume
+    },
+  });
+};
+
+const extensionRequestMicrophonePermission = async (): Promise<MediaStream> => {
+  const state = await getMicrophonePermissionState();
+
+  if (state === "denied") {
+    if (isChromeExtension()) {
+      // Open Chrome extension settings so user can manually re-enable microphone.
+      await chrome.tabs.create({
+        url: `chrome://settings/content/siteDetails?site=chrome-extension%3A%2F%2F${chrome.runtime.id}%2F`,
+      });
+    }
+    // Firefox has no deep link to extension site settings
+    throw new DOMException("Microphone permission denied", "NotAllowedError");
+  }
+
+  if (state === "prompt") {
+    // Side panel can't show the prompt — fall back to popup window
+    await openMicrophoneAccessPopup();
+    const newState = await getMicrophonePermissionState();
+    if (newState !== "granted") {
+      throw new DOMException(
+        "Microphone permission not granted",
+        "NotAllowedError"
+      );
+    }
+    return navigator.mediaDevices.getUserMedia({ audio: true });
+  }
+  // Permission is already granted
+  try {
+    return await navigator.mediaDevices.getUserMedia({ audio: true });
+  } catch (error) {
+    // Firefox bug: when the microphone permission is set to "Always Ask",
+    // navigator.permissions.query incorrectly reports the state as "granted"
+    // instead of "prompt". This causes getUserMedia to fail in the side panel
+    // (which cannot display the native permission prompt). We fall back to the
+    // popup flow so the user can still grant access.
+    if (
+      isFirefoxExtension() &&
+      error instanceof DOMException &&
+      error.name === "NotAllowedError"
+    ) {
+      await openMicrophoneAccessPopup();
+      const newState = await getMicrophonePermissionState();
+      if (newState !== "granted") {
+        throw new DOMException(
+          "Microphone permission not granted",
+          "NotAllowedError"
+        );
+      }
+      return navigator.mediaDevices.getUserMedia({ audio: true });
+    }
+    throw error;
+  }
+};
+
+const getMicrophonePermissionState = async (): Promise<PermissionState> => {
+  const status = await navigator.permissions.query({
+    name: "microphone",
+  });
+  return status.state;
+};
+
+const openMicrophoneAccessPopup = async (): Promise<void> => {
+  const createdWindow = await chrome.windows.create({
+    url: chrome.runtime.getURL("request-mic.html"),
+    type: "popup",
+    width: 400,
+    height: 400,
+  });
+
+  if (!createdWindow?.id) {
+    return;
+  }
+
+  const windowId = createdWindow.id;
+
+  await new Promise<void>((resolve, reject) => {
+    const timeout = setTimeout(() => {
+      chrome.windows.onRemoved.removeListener(onWindowClose);
+      chrome.windows.remove(windowId);
+      reject(new Error("Microphone access popup timed out"));
+    }, MICROPHONE_POPUP_TIMEOUT_MS);
+
+    function onWindowClose(closedWindowId: number) {
+      if (closedWindowId === windowId) {
+        clearTimeout(timeout);
+        chrome.windows.onRemoved.removeListener(onWindowClose);
+        resolve();
+      }
+    }
+
+    chrome.windows.onRemoved.addListener(onWindowClose);
+  });
+};
